Python
Did some cosmetic changes to the Batch file.
Finished the Python version of the program. Was slightly more tricky (to
a degree so minor it's ridiculous).
It's formatted prettier in the Python version since i felt like adding
extra effort on that one once i got into the DateTime module nonsense.